refactor: move Button.tsx into utils
