+++ src/App.scss
@@ -14,11 +14,13 @@ $scrollbar-thumb-color: rgb(123, 123, 123);
         padding-bottom: 12.5vh;
     }
 }
-
+//.swiper-pagination-bullet-color{
 .swiper-pagination-bullet,
 .swiper-pagination-bullet-active {
     background-color: rgb(0,194,203) !important;
 }
+//}
+
 
 // Scrollbar
