+++ src/components/home/Download.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button } from '../Button';
+import { Button } from '../utils/Button';
 
 export const Download: React.FC = () => {
     const urls = {
