+++ src/components/projects/A32NX/Download.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button } from '../../Button';
+import { Button } from '../../utils/Button';
 
 export const Download: React.FC = () => {
     const urls = {
